chore: move context setup outside
