+++ src/context/ResourcesContext.tsx
@@ -1,4 +1,5 @@
 import { createContext } from 'react'
+import { PLAYER_MAX_RESOURCE_CRYSTALS, PLAYER_MAX_RESOURCE_GAS } from '../constants'
 
 type ResourcesContext = {
   crystals: number
@@ -17,3 +18,37 @@ export const ResourcesContext = createContext<ResourcesContext>({
   addGas: () => null,
   removeGas: () => null
 })
+
+export const buildResourcesContext = (
+  crystals: number,
+  setCrystals: React.Dispatch<React.SetStateAction<number>>,
+  gas: number,
+  setGas: React.Dispatch<React.SetStateAction<number>>
+) => {
+  const addCrystals = (val: number) => setCrystals(curr => {
+    const n = curr + val
+    return n > PLAYER_MAX_RESOURCE_CRYSTALS ? PLAYER_MAX_RESOURCE_CRYSTALS : n
+  })
+  const removeCrystals = (val: number) => setCrystals(curr => {
+    const n = curr - val
+    return n < 0 ? 0 : n
+  })
+
+  const addGas = (val: number) => setGas(curr => {
+    const n = curr + val
+    return n > PLAYER_MAX_RESOURCE_GAS ? PLAYER_MAX_RESOURCE_GAS : n
+  })
+  const removeGas = (val: number) => setGas(curr => {
+    const n = curr - val
+    return n < 0 ? 0 : n
+  })
+
+  return {
+    crystals,
+    addCrystals,
+    removeCrystals,
+    gas,
+    addGas,
+    removeGas
+  }
+}
